Updated RenderLib
Moved all renderView functions from ModelLib to RenderLib.
Updated projectedLines, projectedTriangles and projectedQuads to use projectedPoints function in MathLib.
